requirements.txt
Django==3.1.2
djangorestframework==3.12.1
psycopg2-binary==2.8.6
django-filter==2.4.0
djoser==2.1.0
pytest==6.2.4
pytest-django==4.4.0
model-bakery==1.3.2